# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Raffle Draw",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/get-machines": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get Machines",
                "operationId": "b7f31b4adf358c36cd0bdc2c0bd6be14",
                "responses": {
                    "200": {
                        "description": "Get Machines"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/category/winning-tags": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get Machines",
                "operationId": "ddc7f0b7d3fc9025471c6fd1bb740658",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "category id of the winning tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get category id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/winning-tags": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get Machines",
                "operationId": "13446f60965839d7ab395478fd481369",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "category id of the winning tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get category id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agent/winning-tags": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get Machines",
                "operationId": "e02433d475fa5e803a828b10e7c52abf",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "category id of the winning tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get category id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/get-tools": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get Tools",
                "operationId": "5ef535558c825cd86eba8bd3a286db67",
                "responses": {
                    "200": {
                        "description": "Get Tools"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/winning-tags": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get all winning tags",
                "operationId": "e53cfa18d12b9ae9aa4ee7aa76de4726",
                "responses": {
                    "200": {
                        "description": "Get all winning tags"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/winning-tags/create": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "operationId": "2d58826a3f94a2a732f7801efb283aea",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the winning tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "category_id: 1 = Riders, 2 = Machines, 3 = Tools ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stake_price",
                        "in": "query",
                        "description": "stake_price",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "image must be in base64",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}